knowledge-mgt/requirements.txt
# Core dependencies
streamlit==1.31.0
anthropic==0.18.1
chromadb==0.4.22
langchain==0.1.7
langchain-community==0.0.20
langchain-anthropic==0.1.1
openai==1.12.0
python-dotenv==1.0.1

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
# Use latest pandas with wheels for newer Python
pandas==2.2.3
openpyxl==3.1.2

# Embeddings and ML - use latest versions with wheels
sentence-transformers==2.3.1
tiktoken==0.5.2
# Use compatible numpy version with streamlit
numpy==1.26.4
scikit-learn==1.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
watchdog==3.0.0

# Development dependencies
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
pylint==3.0.3
mypy==1.8.0